# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flatmates Shared Expenses API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  clerkId: string;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface Group {
  id: number;
  name: string;
  createdAt: string;
}

export interface Membership {
  id: number;
  groupId: number;
  userId: number;
  userName: string;
  userEmail?: string;
  joinedAt: string;
  /** @nullable */
  leftAt?: string | null;
  isActive: boolean;
}

export interface GroupDetail {
  id: number;
  name: string;
  createdAt: string;
  members: Membership[];
}

export interface GroupInput {
  name: string;
  memberNames?: string[];
}

export interface GroupUpdate {
  name?: string;
}

export interface MembershipInput {
  userId: number;
  joinedAt: string;
  userName?: string;
}

export interface MembershipUpdate {
  /** @nullable */
  leftAt?: string | null;
  joinedAt?: string;
}

export type ExpenseSplitType = typeof ExpenseSplitType[keyof typeof ExpenseSplitType];


export const ExpenseSplitType = {
  equal: 'equal',
  percentage: 'percentage',
  exact: 'exact',
  share: 'share',
  settlement: 'settlement',
} as const;

export interface Expense {
  id: number;
  groupId: number;
  description: string;
  amount: number;
  currency: string;
  amountInr: number;
  splitType: ExpenseSplitType;
  date: string;
  paidByUserId: number;
  paidByName: string;
  /** @nullable */
  notes?: string | null;
  isSettlement?: boolean;
  createdAt: string;
}

export interface ExpenseSplit {
  id: number;
  expenseId: number;
  userId: number;
  userName: string;
  amountOwed: number;
  /** @nullable */
  shareCount?: number | null;
  /** @nullable */
  percentage?: number | null;
}

export interface ExpenseDetail {
  id: number;
  groupId: number;
  description: string;
  amount: number;
  currency: string;
  amountInr: number;
  splitType: string;
  date: string;
  paidByUserId: number;
  paidByName: string;
  /** @nullable */
  notes?: string | null;
  isSettlement?: boolean;
  createdAt: string;
  splits: ExpenseSplit[];
}

export type ExpenseInputSplitType = typeof ExpenseInputSplitType[keyof typeof ExpenseInputSplitType];


export const ExpenseInputSplitType = {
  equal: 'equal',
  percentage: 'percentage',
  exact: 'exact',
  share: 'share',
  settlement: 'settlement',
} as const;

export interface SplitInput {
  userId: number;
  amountOwed: number;
  shareCount?: number;
  percentage?: number;
}

export interface ExpenseInput {
  description: string;
  amount: number;
  currency: string;
  splitType: ExpenseInputSplitType;
  date: string;
  paidByUserId: number;
  notes?: string;
  isSettlement?: boolean;
  splits?: SplitInput[];
}

export interface ExpenseUpdate {
  description?: string;
  amount?: number;
  currency?: string;
  splitType?: string;
  date?: string;
  paidByUserId?: number;
  notes?: string;
  splits?: SplitInput[];
}

export interface Settlement {
  id: number;
  groupId: number;
  fromUserId: number;
  fromUserName: string;
  toUserId: number;
  toUserName: string;
  amount: number;
  date: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface SettlementInput {
  fromUserId: number;
  toUserId: number;
  amount: number;
  date: string;
  notes?: string;
}

export interface MemberBalance {
  userId: number;
  userName: string;
  totalPaid: number;
  totalOwed: number;
  netBalance: number;
}

export interface BalanceSummary {
  members: MemberBalance[];
  totalSpend: number;
  currency: string;
}

export interface SuggestedSettlement {
  fromUserId: number;
  fromUserName: string;
  toUserId: number;
  toUserName: string;
  amount: number;
}

export interface MemberStat {
  userId: number;
  userName: string;
  totalPaid: number;
  shareOwed: number;
  netBalance: number;
}

export interface GroupStats {
  totalSpend: number;
  totalSettled: number;
  outstandingDebt: number;
  expenseCount: number;
  memberStats: MemberStat[];
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  expense: 'expense',
  settlement: 'settlement',
  member_joined: 'member_joined',
  member_left: 'member_left',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  description: string;
  date: string;
  userName: string;
  /** @nullable */
  amount: number | null;
  /** @nullable */
  currency?: string | null;
}

export interface ExchangeRate {
  id: number;
  fromCurrency: string;
  toCurrency: string;
  rate: number;
  effectiveDate: string;
}

export interface ExchangeRateInput {
  fromCurrency: string;
  toCurrency: string;
  rate: number;
  effectiveDate: string;
}

export interface ImportFileInput {
  /** Raw CSV file content as a string */
  csvContent: string;
  groupId: number;
}

export type ImportIssueSeverity = typeof ImportIssueSeverity[keyof typeof ImportIssueSeverity];


export const ImportIssueSeverity = {
  error: 'error',
  warning: 'warning',
  info: 'info',
} as const;

export interface ImportIssue {
  rowNumber: number;
  issueType: string;
  description: string;
  actionTaken: string;
  severity: ImportIssueSeverity;
  /** @nullable */
  rawRow?: string | null;
}

export type ImportPreviewRowStatus = typeof ImportPreviewRowStatus[keyof typeof ImportPreviewRowStatus];


export const ImportPreviewRowStatus = {
  ok: 'ok',
  skipped: 'skipped',
  needs_review: 'needs_review',
} as const;

export interface ImportPreviewRow {
  rowNumber: number;
  description: string;
  /** @nullable */
  amount: number | null;
  currency: string;
  paidBy: string;
  splitType: string;
  date: string;
  status: ImportPreviewRowStatus;
  issues?: ImportIssue[];
}

export interface ImportPreview {
  sessionToken: string;
  totalRows: number;
  validRows: number;
  skippedRows: number;
  issueCount: number;
  issues: ImportIssue[];
  rows: ImportPreviewRow[];
}

export interface ImportConfirmInput {
  sessionToken: string;
  groupId: number;
}

export interface ImportResult {
  sessionId: number;
  importedCount: number;
  skippedCount: number;
  issueCount: number;
}

export interface ImportSession {
  id: number;
  filename: string;
  groupId: number;
  importedAt: string;
  importedCount: number;
  skippedCount: number;
  issueCount: number;
}

export interface ImportSessionDetail {
  id: number;
  filename: string;
  groupId: number;
  importedAt: string;
  importedCount: number;
  skippedCount: number;
  issues: ImportIssue[];
}