# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Flatmates Shared Expenses API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups/{groupId}/expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses for a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expenses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense in a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{expenseId}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get expense with splits
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseDetail"
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Settlements ────────────────────────────────────────────────────────────
...
  /groups/{groupId}/balances:
    get:
      operationId: getGroupBalances
      tags: [balances]
      summary: Get per-member balances (net owed/owes)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Balances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceSummary"
  /groups/{groupId}/settlements/suggested:
    get:
      operationId: getSuggestedSettlements
      tags: [balances]
      summary: Get minimal set of payments to settle all debts
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suggested settlements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SuggestedSettlement"
  /groups/{groupId}/stats:
    get:
      operationId: getGroupStats
      tags: [balances]
      summary: Group spending stats and per-member totals
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupStats"
  /groups/{groupId}/activity:
    get:
      operationId: getGroupActivity
      tags: [balances]
      summary: Recent activity timeline for a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  # ── Exchange Rates ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Expense:
      type: object
      required: [id, groupId, description, amount, currency, amountInr, splitType, date, paidByUserId, paidByName, createdAt]
      properties:
        id:
          type: integer
        groupId:
          type: integer
        description:
          type: string
        amount:
          type: number
        currency:
          type: string
        amountInr:
          type: number
        splitType:
          type: string
          enum: [equal, percentage, exact, share, settlement]
        date:
          type: string
        paidByUserId:
          type: integer
        paidByName:
          type: string
        notes:
          type: ["string", "null"]
        isSettlement:
          type: boolean
        createdAt:
          type: string
    ExpenseDetail:
      type: object
      required: [id, groupId, description, amount, currency, amountInr, splitType, date, paidByUserId, paidByName, createdAt, splits]
      properties:
        id:
          type: integer
        groupId:
          type: integer
        description:
          type: string
        amount:
          type: number
        currency:
          type: string
        amountInr:
          type: number
        splitType:
          type: string
        date:
          type: string
        paidByUserId:
          type: integer
        paidByName:
          type: string
        notes:
          type: ["string", "null"]
        isSettlement:
          type: boolean
        createdAt:
          type: string
        splits:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseSplit"
    ExpenseInput:
      type: object
      required: [description, amount, currency, splitType, date, paidByUserId]
      properties:
        description:
          type: string
        amount:
          type: number
        currency:
          type: string
        splitType:
          type: string
          enum: [equal, percentage, exact, share, settlement]
        date:
          type: string
        paidByUserId:
          type: integer
        notes:
          type: string
        isSettlement:
          type: boolean
        splits:
          type: array
          items:
            $ref: "#/components/schemas/SplitInput"
    ExpenseUpdate:
      type: object
      properties:
        description:
          type: string
        amount:
          type: number
        currency:
          type: string
        splitType:
          type: string
        date:
          type: string
        paidByUserId:
          type: integer
        notes:
          type: string
        splits:
          type: array
          items:
            $ref: "#/components/schemas/SplitInput"
    ExpenseSplit:
      type: object
      required: [id, expenseId, userId, userName, amountOwed]
      properties:
        id:
          type: integer
        expenseId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        amountOwed:
          type: number
        shareCount:
          type: ["number", "null"]
        percentage:
          type: ["number", "null"]
    SplitInput:
      type: object
      required: [userId, amountOwed]
      properties:
        userId:
          type: integer
        amountOwed:
          type: number
        shareCount:
          type: number
        percentage:
          type: number

    # ── Settlements ───────────────────────────────────────────────────────────
...
    MemberBalance:
      type: object
      required: [userId, userName, totalPaid, totalOwed, netBalance]
      properties:
        userId:
          type: integer
        userName:
          type: string
        totalPaid:
          type: number
        totalOwed:
          type: number
        netBalance:
          type: number
    BalanceSummary:
      type: object
      required: [members, totalSpend, currency]
      properties:
        members:
          type: array
          items:
            $ref: "#/components/schemas/MemberBalance"
        totalSpend:
          type: number
        currency:
          type: string
    SuggestedSettlement:
      type: object
      required: [fromUserId, fromUserName, toUserId, toUserName, amount]
      properties:
        fromUserId:
          type: integer
        fromUserName:
          type: string
        toUserId:
          type: integer
        toUserName:
          type: string
        amount:
          type: number
    GroupStats:
      type: object
      required: [totalSpend, totalSettled, outstandingDebt, expenseCount, memberStats]
      properties:
        totalSpend:
          type: number
        totalSettled:
          type: number
        outstandingDebt:
          type: number
        expenseCount:
          type: integer
        memberStats:
          type: array
          items:
            $ref: "#/components/schemas/MemberStat"
    MemberStat:
      type: object
      required: [userId, userName, totalPaid, shareOwed, netBalance]
      properties:
        userId:
          type: integer
        userName:
          type: string
        totalPaid:
          type: number
        shareOwed:
          type: number
        netBalance:
          type: number
    ActivityItem:
      type: object
      required: [id, type, description, date, userName, amount]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [expense, settlement, member_joined, member_left]
        description:
          type: string
        date:
          type: string
        userName:
          type: string
        amount:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]

    # ── Exchange Rates ─────────────────────────────────────────────────────────
...