# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flatmates Shared Expenses API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemType } from './activityItemType';

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  description: string;
  date: string;
  userName: string;
  /** @nullable */
  amount: number | null;
  /** @nullable */
  currency?: string | null;
}