# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flatmates Shared Expenses API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseSplitType } from './expenseSplitType';

export interface Expense {
  id: number;
  groupId: number;
  description: string;
  amount: number;
  currency: string;
  amountInr: number;
  splitType: ExpenseSplitType;
  date: string;
  paidByUserId: number;
  paidByName: string;
  /** @nullable */
  notes?: string | null;
  isSettlement?: boolean;
  createdAt: string;
}