# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flatmates Shared Expenses API
 * OpenAPI spec version: 0.1.0
 */
import type { SplitInput } from './splitInput';

export interface ExpenseUpdate {
  description?: string;
  amount?: number;
  currency?: string;
  splitType?: string;
  date?: string;
  paidByUserId?: number;
  notes?: string;
  splits?: SplitInput[];
}