# lib/api-zod/src/generated/types/memberStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flatmates Shared Expenses API
 * OpenAPI spec version: 0.1.0
 */

export interface MemberStat {
  userId: number;
  userName: string;
  totalPaid: number;
  shareOwed: number;
  netBalance: number;
}